+++ tests/fido_otp_container_test.c
@@ -155,7 +155,7 @@ int file_put_data(file_t *file, const_byte_array_t data) {
         return PICOKEYS_ERR_NO_MEMORY;
     }
     memcpy(test_file->storage, data.data, data.len);
-    test_file->size = data.len;
+    test_file->size = (uint32_t)data.len;
     test_file->file.data = data.len > 0 ? test_file->storage : NULL;
     return PICOKEYS_OK;
 }
